feat: rename webconnectivity v0.5 as webconnectivitylte
In this way, we can avoid a package name conflict when moving
the v0.4 webconnectivity inside internal/experiment.

Part of https://github.com/ooni/probe/issues/2115
