+++ internal/experiment/webconnectivitylte/analysiscore.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 import (
 	"fmt"
