+++ internal/experiment/webconnectivitylte/analysisdns.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // DNS analysis
