+++ internal/experiment/webconnectivitylte/analysishttpcore.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // HTTP core analysis
