+++ internal/experiment/webconnectivitylte/analysishttpdiff.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // HTTP diff analysis
