+++ internal/experiment/webconnectivitylte/analysistcpip.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // TCP/IP analysis
