+++ internal/experiment/webconnectivitylte/analysistls.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // TLS analysis
