+++ internal/experiment/webconnectivitylte/cleartextflow.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // CleartextFlow
