+++ internal/experiment/webconnectivitylte/dnscache.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 import "sync"
