+++ internal/experiment/webconnectivitylte/dnsresolvers.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // DNSResolvers
