+++ internal/experiment/webconnectivitylte/dnswhoami.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 import (
 	"context"
