+++ internal/experiment/webconnectivitylte/doc.go
@@ -1,7 +1,7 @@
-// Package webconnectivity implements the web_connectivity experiment.
+// Package webconnectivitylte implements the web_connectivity experiment.
 //
 // Spec: https://github.com/ooni/spec/blob/master/nettests/ts-017-web-connectivity.md.
 //
 // This implementation, in particular, contains extensions over the original model,
 // which we document at https://github.com/ooni/probe/issues/2237.
-package webconnectivity
+package webconnectivitylte
