+++ internal/experiment/webconnectivitylte/inputparser.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // Input parsing
