+++ internal/experiment/webconnectivitylte/measurer.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // Measurer
