+++ internal/experiment/webconnectivitylte/secureflow.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // SecureFlow
