+++ internal/experiment/webconnectivitylte/summary.go
@@ -1,4 +1,4 @@
-package webconnectivity
+package webconnectivitylte
 
 //
 // Summary
